+++ tests/testthat/test_flagsCount.R
@@ -1,0 +1,11 @@
+## library(devtools)
+## load_all()
+
+context("flagsCount")
+
+test_that("No flags table provided",{
+
+    pk <- readRDS(file=system.file("examples/data/xgxr2.rds",package="NMdata"))
+
+    flagsCount(pk[EVID==0])
+})