+++ tests/testthat/test_fnAppend.R
@@ -19,10 +19,20 @@ test_that("basic",{
     
 })
 
+## library(devtools)
+## load_all("~/wdirs/NMdata")
 test_that("skip directory double dots",{
     ## todo. This should return an error. There is no extension to
     ## append in front of.
-    fnAppend("fe/../egef","hmm")
+    expect_error(fnAppend("fe/../egef","hmm"))
     ## should also return error:
-    fnAppend("egef","hmm")
+    expect_error(fnAppend("egef","hmm"))
+})
+
+test_that("empty string does notning",{
+
+    str1 <- "fe.ef"
+    str2 <- fnAppend(str1,"")
+    expect_identical(str1,str2)
+
 })